Color pallete file renamed
